# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Neo FMC - Multi-tenant Microfinance SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientListResponse"
    post:
      operationId: createClient
      tags: [clients]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientResponse"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientResponse"
    put:
      operationId: updateClient
      tags: [clients]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ClientResponse:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        nationalId:
          type: string
        fullNameAr:
          type: string
        fullNameEn:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        riskScore:
          type: integer
          nullable: true
        isBlacklisted:
          type: boolean
        blacklistReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, nationalId, fullNameAr, isBlacklisted, createdAt]
    CreateClientRequest:
      type: object
      properties:
        nationalId:
          type: string
        fullNameAr:
          type: string
        fullNameEn:
          type: string
        phone:
          type: string
        address:
          type: string
        riskScore:
          type: integer
      required: [nationalId, fullNameAr]
    ClientListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ClientResponse"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
...